// config/enum.ts
/* eslint-disable no-unused-vars */
export enum FormFieldTypes {
  INPUT = "input",
  TEXTAREA = "textarea",
  PHONE_INPUT = "phoneInput",
  PASSWORD = "password",
  CHECKBOX = "checkbox",
  DATE_PICKER = "datePicker",
  SELECT = "select",
  SKELETON = "skeleton",
  RADIO_GROUP = "radioGroup",
  SWITCH = "switch",
  RATING = "rating",
  TOGGLE_GROUP = "toggleGroup",
  COMBOBOX = "combobox",
  FILE_INPUT = "fileInput",
  MULTI_SELECT = "multiSelect",
  COMMAND = "command",
  PRICE_INPUT = "priceInput",
  QUILL = "quill",
  INPT_OTP = "OTP",
}